feat: Complete project structure refactoring
🎯 MAJOR REFACTORING: Consolidated project structure from scattered 11-directory flat layout into logical hierarchical organization

📁 NEW STRUCTURE:
- src/core/         : System fundamentals & lifecycle (9 files)
- src/hardware/     : Clean hardware abstraction (4 files)
- src/display/      : Display hardware & LVGL platform (4 files)
- src/messaging/    : Organized messaging system (11 files in 4 subdirs)
- src/ota/          : CONSOLIDATED OTA system (6 files, resolved duplicates)
- src/logo/         : UNIFIED logo system (9 files, services moved here)
- src/application/  : Business logic services (7 files in 2 subdirs)
- src/ui/           : SquareLine Studio Generated (untouched)

🔧 CONSOLIDATIONS:
- OTA: Merged duplicate OTA managers from boot/ + hardware/ → ota/
- Logo: Unified logo system, moved services from application/services/ → logo/
- Core: Consolidated system lifecycle from 5 scattered directories → core/

📊 IMPACT:
- Total files: 62 source files successfully reorganized
- Directories: Reduced from 11 flat to 8 logical hierarchies
- Include paths: Comprehensively updated for all relative path variations
- Empty directories: Automatically cleaned up

✅ VERIFIED:
- All critical files relocated correctly
- Include paths systematically updated
- Audio system path issues resolved
- Comprehensive test environment validation passed

🧹 AUTOMATION:
- Complete scripted refactoring system created
- Isolated test environment validation
- Automated include path resolution
- Safety checks and verification at every step
