+++ src/core/BootManager.cpp
