+++ src/core/CoreLoggingFilter.cpp
