+++ src/core/TaskManager.h
