+++ src/logo/LogoSupplier.cpp
