+++ src/logo/LogoSupplier.h
