+++ src/logo/MessageBusLogoSupplier.cpp
