+++ src/logo/MessageBusLogoSupplier.h
